Wrap connection instances to provide logging or other accessory features.
The main benefit of this approach is that logging can now be used on all the
classes implementing the Predis\Network\IConnectionSingle interface such as
Predis\Network\WebdisConnection, but also custom classes implemented by users.

Another plus point is in the lower overhead in the vast majority of cases
since the wrapper class is actually used only when logging is enabled for
specific connections.