Specify the password via parameters for non sentinel connections
Before this change, the options:parameters:password was reserved for sentinel
connections only, as we need a way to pass the auth to the servers that
sentinel provides for us, but not to sentinel itself.

This change allows us to use options:parameters:password for regular redis
connections.

The code checks if replication is set, because we want to keep the sentinel
functionality as is, which in this case is to leave the password in parameters
and allow sentinel to build the connection later.