Authenticate against Redis using a password
Redis 6 will throw a "ERR AUTH called without any password configured for the
default user" error when we try to authenticate with a password when no password
is required.

By requiring a password (and using it when connecting with Redis) we prevent
this error.

See https://github.com/snc/SncRedisBundle/pull/599#issuecomment-751125874 for
more information.